projects/mapforge/requirements.txt
# MapForge dependencies

# Official PixelLab client (v1 API)
pixellab>=0.1.0

# HTTP requests (for v2 tileset API)
requests>=2.28.0

# Image processing
Pillow>=9.0.0

# Environment variable loading
python-dotenv>=1.0.0
